+++ signup.html
@@ -6,7 +6,7 @@
         <title>Linkbrary signup</title>
         <link href="./styles/reset.css" rel="stylesheet" type="text/css" />
         <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css" />
-        <link href="./styles/signin.css" rel="stylesheet" type="text/css" />
+        <link href="./styles/signup.css" rel="stylesheet" type="text/css" />
     </head>
     <body>
         <div class="root">
@@ -46,12 +46,9 @@
                       비밀번호 확인
                   </label>
                   <div class="input-password">
-                    <div class="input-password-wrapper">
+                    <div class="input-password-wrapper2">
                       <input type="password" placeholder="비밀번호를 한번 더 입력해주세요" id="password-check-label" name="password-checkLabel">
-                      <img src="./assets/png/eye-off.png" alt="click to see password" class="password-check-eye-off" style="position: absolute;
-                      top: 50%;
-                      right: 1%;
-                      transform: translate(-50%, -50%);">
+                      <img src="./assets/png/eye-off.png" alt="click to see password" class="password-check-eye-off">
                     </div>
                   </div>
                 </div>
